added ProductRepository, db scripts, changed products.html template
