+++ src/main/java/com/example/internship/repository/ProductRepository.java
@@ -1,0 +1,7 @@
+package com.example.internship.repository;
+
+import com.example.internship.entity.Product;
+import org.springframework.data.repository.CrudRepository;
+
+public interface ProductRepository extends CrudRepository<Product, Long> {
+}
