+++ src/main/resources/db/changelog/2020-08-26-01/01-insert-products.yaml
@@ -1,0 +1,69 @@
+databaseChangeLog:
+- changeSet:
+    id: 2020-08-26
+    author: Murashov Alexey
+    comment: Inserting some products
+    changes:
+      - insert:
+          tableName: product
+          columns:
+            - column:
+                name: id
+                value: "1"
+            - column:
+                name: category_id
+                value: "1"
+            - column:
+                name: name
+                value: "IPhone 7"
+            - column:
+                name: description
+                value: "IPhone 7 description"
+            - column:
+                name: picture
+                value: ""
+            - column:
+                name: status_id
+                value: "1"
+      - insert:
+          tableName: product
+          columns:
+            - column:
+                name: id
+                value: "2"
+            - column:
+                name: category_id
+                value: "1"
+            - column:
+                name: name
+                value: "IPhone 8"
+            - column:
+                name: description
+                value: "IPhone 8 description"
+            - column:
+                name: picture
+                value: ""
+            - column:
+                name: status_id
+                value: "1"
+      - insert:
+          tableName: product
+          columns:
+            - column:
+                name: id
+                value: "3"
+            - column:
+                name: category_id
+                value: "1"
+            - column:
+                name: name
+                value: "IPhone 10"
+            - column:
+                name: description
+                value: "IPhone 10 description"
+            - column:
+                name: picture
+                value: ""
+            - column:
+                name: status_id
+                value: "2"
