+++ src/main/resources/db/changelog/2020-08-26-01/02-insert-product-statuses.yaml
@@ -1,0 +1,24 @@
+databaseChangeLog:
+  - changeSet:
+      id: 2020-08-26
+      author: Murashov Alexey
+      comment: Inserting some statuses
+      changes:
+        - insert:
+            tableName: product_status
+            columns:
+              - column:
+                  name: id
+                  value: "1"
+              - column:
+                  name: description
+                  value: "В наличии"
+        - insert:
+            tableName: product_status
+            columns:
+              - column:
+                  name: id
+                  value: "2"
+              - column:
+                  name: description
+                  value: "Нет в наличии"
