+++ src/main/resources/db/changelog/2020-08-26-01/03-insert-categories.yaml
@@ -1,0 +1,24 @@
+databaseChangeLog:
+  - changeSet:
+      id: 2020-08-26
+      author: Murashov Alexey
+      comment: Inserting some statuses
+      changes:
+        - insert:
+            tableName: category
+            columns:
+              - column:
+                  name: id
+                  value: "1"
+              - column:
+                  name: name
+                  value: "Смартфоны"
+        - insert:
+            tableName: category
+            columns:
+              - column:
+                  name: id
+                  value: "2"
+              - column:
+                  name: name
+                  value: "Планшеты"
