+++ src/main/resources/db/changelog/db.changelog-master.yaml
@@ -3,3 +3,5 @@ databaseChangeLog:
       file: db/changelog/changes/v_0.0.1.yaml
   - include:
       file: db/changelog/changes/v_0.0.2.yaml
+  - include:
+      file: db/changelog/changes/v_0.0.3.yaml
